+++ src/redturtle/prenotazioni/behaviors/booking_folder/notification_sms/events.py
@@ -7,10 +7,19 @@ from zope.globalrequest import getRequest
 from redturtle.prenotazioni.interfaces import IBookingNotificationSender
 from redturtle.prenotazioni.interfaces import IPrenotazioneSMSMessage
 
+from .notification_sms import INotificationSMS
+
 logger = getLogger(__name__)
 
 
+def booking_folder_provides_current_behavior(booking):
+    return INotificationSMS.providedBy(booking.getPrenotazioniFolder())
+
+
 def send_notification_on_transition(context, event) -> None:
+    if not booking_folder_provides_current_behavior(context):
+        return
+
     booking_folder = context.getPrenotazioniFolder()
     flags = {
         i: getattr(booking_folder, f"notify_on_{i}", False)
@@ -45,6 +54,9 @@ def send_notification_on_transition(context, event) -> None:
 
 # TODO: use the notify_on_after_transition_event method techique instead
 def notify_on_move(booking, event):
+    if not booking_folder_provides_current_behavior(booking):
+        return
+
     if not getattr(booking.getPrenotazioniFolder(), "notify_on_move", False):
         return
     if not getattr(booking, "email", ""):
@@ -61,6 +73,9 @@ def notify_on_move(booking, event):
 
 
 def send_booking_reminder(context, event):
+    if not booking_folder_provides_current_behavior(context):
+        return
+
     message_adapter = getMultiAdapter(
         (context, event),
         IPrenotazioneSMSMessage,