+++ src/redturtle/prenotazioni/behaviors/booking_folder/notification_sms/notification_sms.py
@@ -130,16 +130,6 @@ class INotificationSMS(model.Schema):
         max_length=70,
     )
 
-    notify_as_reminder_sms_message = schema.Text(
-        title=_(
-            "notify_as_reminder_sms_message",
-            default="Booking reminder message.",
-        ),
-        description=_("notify_as_reminder_sms_message_help", default=""),
-        defaultFactory=notify_as_reminder_sms_message_default_factory,
-        required=False,
-    )
-
     model.fieldset(
         "SMS Notifications Gateway",
         label=_(