+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -488,6 +488,14 @@ class IPrenotazioniFolder(model.Schema):
     )
 
     model.fieldset(
+        "Reminders",
+        label=_("reminders_label", default="Reminders"),
+        fields=[
+            "app_io_enabled",
+        ],
+    )
+
+    model.fieldset(
         "Notifications",
         label=_("notifications_label", default="Notifications"),
         fields=[
@@ -496,6 +504,9 @@ class IPrenotazioniFolder(model.Schema):
             "notify_on_move",
             "notify_on_refuse",
         ],
+        description=_(
+            "notifications_help", default="Notifications management controlpanel"
+        ),
     )